Move date utilities from platform package to core.date
DateConstants and DateUtils are pure utilities without expect/actual
declarations. Per package organization rules, only expect/actual code
should be in the *.platform.* package. Moved to core.date package
within the same module.
