+++ core/platform/src/commonMain/kotlin/space/be1ski/vibits/core/date/DateUtils.kt
@@ -1,4 +1,4 @@
-package space.be1ski.vibits.core.platform.date
+package space.be1ski.vibits.core.date
 
 import kotlinx.datetime.DatePeriod
 import kotlinx.datetime.DayOfWeek
@@ -6,9 +6,6 @@ import kotlinx.datetime.LocalDate
 import kotlinx.datetime.Month
 import kotlinx.datetime.minus
 
-/**
- * Returns the Monday of the week containing [date].
- */
 fun startOfWeek(date: LocalDate): LocalDate {
   var start = date
   while (start.dayOfWeek != DayOfWeek.MONDAY) {
@@ -17,12 +14,6 @@ fun startOfWeek(date: LocalDate): LocalDate {
   return start
 }
 
-/**
- * Returns the quarter index (1-4) for a date.
- */
 fun quarterIndex(date: LocalDate): Int = quarterIndex(date.month)
 
-/**
- * Returns the quarter index (1-4) for a month.
- */
 fun quarterIndex(month: Month): Int = month.ordinal / MONTHS_IN_QUARTER + FIRST_QUARTER_INDEX
