+++ feature/habits/domain/src/commonMain/kotlin/space/be1ski/vibits/feature/habits/domain/usecase/GenerateActivityRangesUseCase.kt
@@ -3,12 +3,12 @@ package space.be1ski.vibits.feature.habits.domain.usecase
 import kotlinx.datetime.DatePeriod
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.plus
-import space.be1ski.vibits.core.platform.date.DAYS_IN_WEEK
-import space.be1ski.vibits.core.platform.date.FIRST_DAY_OF_MONTH
-import space.be1ski.vibits.core.platform.date.FIRST_QUARTER_INDEX
-import space.be1ski.vibits.core.platform.date.QUARTERS_IN_YEAR
-import space.be1ski.vibits.core.platform.date.quarterIndex
-import space.be1ski.vibits.core.platform.date.startOfWeek
+import space.be1ski.vibits.core.date.DAYS_IN_WEEK
+import space.be1ski.vibits.core.date.FIRST_DAY_OF_MONTH
+import space.be1ski.vibits.core.date.FIRST_QUARTER_INDEX
+import space.be1ski.vibits.core.date.QUARTERS_IN_YEAR
+import space.be1ski.vibits.core.date.quarterIndex
+import space.be1ski.vibits.core.date.startOfWeek
 import space.be1ski.vibits.feature.habits.domain.model.ActivityRange
 
 object GenerateActivityRangesUseCase {
