+++ feature/habits/domain/src/commonMain/kotlin/space/be1ski/vibits/feature/habits/domain/usecase/NavigateActivityRangeUseCase.kt
@@ -3,8 +3,8 @@ package space.be1ski.vibits.feature.habits.domain.usecase
 import kotlinx.datetime.DatePeriod
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.plus
-import space.be1ski.vibits.core.platform.date.DAYS_IN_WEEK
-import space.be1ski.vibits.core.platform.date.QUARTERS_IN_YEAR
+import space.be1ski.vibits.core.date.DAYS_IN_WEEK
+import space.be1ski.vibits.core.date.QUARTERS_IN_YEAR
 import space.be1ski.vibits.feature.habits.domain.model.ActivityRange
 
 /**
