+++ feature/habits/domain/src/commonTest/kotlin/space/be1ski/vibits/feature/habits/domain/usecase/DateCalculationsUseCaseTest.kt
@@ -4,8 +4,8 @@ import kotlinx.datetime.DayOfWeek
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.Month
 import kotlinx.datetime.TimeZone
-import space.be1ski.vibits.core.platform.date.quarterIndex
-import space.be1ski.vibits.core.platform.date.startOfWeek
+import space.be1ski.vibits.core.date.quarterIndex
+import space.be1ski.vibits.core.date.startOfWeek
 import space.be1ski.vibits.feature.memos.domain.model.Memo
 import kotlin.test.Test
 import kotlin.test.assertEquals
